requirements.txt
tabula-py
tabulate